refactor: loadUsers and move to separate file
